requirements.txt
requests
pandas
pyyaml
python-dotenv
reportlab
openai